Refactor: ProductDetail 컴포넌트 점진적 향상 기법을 이용한 Webp 이미지 적용
